+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 27
     buildToolsVersion "27.0.3"
     defaultConfig {
-        applicationId "kres.realtimeshoppinglist"
+        applicationId "kres.listyMcListFace"
         minSdkVersion 15
         targetSdkVersion 27
         versionCode 1
