+++ app/src/main/java/kres/listyMcListFace/dialog/joinCreate/ShoppingListUtil.java
@@ -1,0 +1,11 @@
+package kres.listyMcListFace.dialog.joinCreate;
+
+import kres.listyMcListFace.firebase.shoppingList.ShoppingListAdapter;
+import kres.listyMcListFace.persistence.PersistenceManager;
+
+public interface ShoppingListUtil {
+
+    ShoppingListAdapter getAdapter();
+    PersistenceManager getPersistenceManger();
+
+}
