+++ app/src/main/java/kres/listyMcListFace/dialog/newProduct/NewProductDialogListener.java
@@ -1,0 +1,10 @@
+package kres.listyMcListFace.dialog.newProduct;
+
+import kres.listyMcListFace.model.Product;
+
+public interface NewProductDialogListener {
+
+    void onProductAdded(Product product);
+    void onCancel();
+
+}
