+++ app/src/main/java/kres/listyMcListFace/firebase/shoppingList/ListExistsListener.java
@@ -1,0 +1,11 @@
+package kres.listyMcListFace.firebase.shoppingList;
+
+import kres.listyMcListFace.model.ShoppingList;
+
+public interface ListExistsListener {
+
+    void onListFound(ShoppingList list);
+
+    void onListNotFound();
+
+}
