+++ app/src/main/java/kres/listyMcListFace/util/Constants.java
@@ -1,0 +1,8 @@
+package kres.listyMcListFace.util;
+
+public class Constants {
+
+    public static final String SHOPPING_LIST_INTENT_KEY = "SHOPPING_LIST";
+
+
+}
