--- app/src/main/java/kres/realtimeshoppinglist/dialog/joinCreate/ShoppingListUtil.java
@@ -1,11 +1,0 @@
-package kres.realtimeshoppinglist.dialog.joinCreate;
-
-import kres.realtimeshoppinglist.firebase.shoppingList.ShoppingListAdapter;
-import kres.realtimeshoppinglist.persistence.PersistenceManager;
-
-public interface ShoppingListUtil {
-
-    ShoppingListAdapter getAdapter();
-    PersistenceManager getPersistenceManger();
-
-}
