--- app/src/main/java/kres/realtimeshoppinglist/dialog/newProduct/NewProductDialogListener.java
@@ -1,10 +1,0 @@
-package kres.realtimeshoppinglist.dialog.newProduct;
-
-import kres.realtimeshoppinglist.model.Product;
-
-public interface NewProductDialogListener {
-
-    void onProductAdded(Product product);
-    void onCancel();
-
-}
