--- app/src/main/java/kres/realtimeshoppinglist/firebase/shoppingList/ListExistsListener.java
@@ -1,11 +1,0 @@
-package kres.realtimeshoppinglist.firebase.shoppingList;
-
-import kres.realtimeshoppinglist.model.ShoppingList;
-
-public interface ListExistsListener {
-
-    void onListFound(ShoppingList list);
-
-    void onListNotFound();
-
-}
