--- app/src/main/java/kres/realtimeshoppinglist/firebase/util/ListID.java
@@ -1,19 +1,0 @@
-package kres.realtimeshoppinglist.firebase.util;
-
-import java.math.BigInteger;
-import java.security.SecureRandom;
-
-public class ListID {
-
-    public static String generate() {
-        SecureRandom random = new SecureRandom();
-        String fullRandomString = new BigInteger(32, random).toString(32);
-
-        if (fullRandomString.length() >= 6) {
-            fullRandomString = fullRandomString.substring(0, 6);
-        }
-
-        return fullRandomString.toUpperCase();
-    }
-
-}
