--- app/src/main/java/kres/realtimeshoppinglist/util/Constants.java
@@ -1,8 +1,0 @@
-package kres.realtimeshoppinglist.util;
-
-public class Constants {
-
-    public static final String SHOPPING_LIST_INTENT_KEY = "SHOPPING_LIST";
-
-
-}
