+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Real Time Shopping List</string>
+    <string name="app_name">Listy McListFace</string>
     <string name="list_code">List Code</string>
     <string name="submit">submit</string>
     <string name="zero">0</string>
